refactor: getResult 함수 WinningLotto로 이동
